Add explicit int/long cast in TypedSet
This silences a warning from IntelliJ, (and the re-wrapping of the
line makes sure you can see the whole expression in a reasonable-sized
window).

+++ core/trino-main/src/main/java/io/trino/operator/aggregation/TypedSet.java
@@ -176,7 +176,10 @@ public class TypedSet
 
     public long getRetainedSizeInBytes()
     {
-        return INSTANCE_SIZE + INT_ARRAY_LIST_INSTANCE_SIZE + elementBlock.getRetainedSizeInBytes() + blockPositionByHash.size() * Integer.BYTES;
+        return INSTANCE_SIZE
+                + INT_ARRAY_LIST_INSTANCE_SIZE
+                + elementBlock.getRetainedSizeInBytes()
+                + blockPositionByHash.size() * (long) Integer.BYTES;
     }
 
     public boolean contains(Block block, int position)
